+++ docs/about/release-notes.rst
@@ -24,6 +24,7 @@ Breaking changes
 * The ``unit`` attribute now distinguishes ``None`` from ``dimensionless``. For number-like dtypes the default is ``dimensionless`` (unchanged) whereas all other dtypes such as strings default to ``None`` (new) `#2395 <https://github.com/scipp/scipp/pull/2395>`_.
 * :py:func:`scipp.spatial.rotations_from_rotvecs` has been changed to taking a variable input rather than separate values/dims/units.
 * ``scipp.NotFoundFound`` error has been replaced with ``KeyError`` in Python `#2416 <https://github.com/scipp/scipp/pull/2416>`_.
+* Some functions now raise ``scipp.DimensionError`` instead of ``scipp.NotFoundFound`` `#2416 <https://github.com/scipp/scipp/pull/2416>`_.
 
 Bugfixes
 ~~~~~~~~